requirements.txt
bs4
python-dotenv
requests
langchain
langchain_community
langchain-groq
streamlit
ipython